+++ backend/src/main/java/com/sudoleg/portfoliomanager/repositories/UserRepository.java
@@ -1,17 +1,18 @@
 package com.sudoleg.portfoliomanager.repositories;
 
-import com.sudoleg.portfoliomanager.domain.entities.UserEntity;
+import java.util.Optional;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
+import com.sudoleg.portfoliomanager.domain.entities.UserEntity;
 
 /**
  * Repository for users.
  * Repositories are used to perform database operations automatically.
  */
 @Repository
-public interface UserRepository extends CrudRepository<UserEntity, Integer> {
+public interface UserRepository extends CrudRepository<UserEntity, Long> {
 
     Optional<UserEntity> findByUsername(String username);
